Runtime Optimization through Axios.All
Requests are processed in batches of 10000 players, or 100 URLs requests. I attempted to do it all requests simultaneously but the Devil Dagger server was not happy with that and rejected my call. Could be optimized further, and I have been challenged to reduce the line count further although with my current implementation it would be possible but uglier. Still a WIP